My Project work (EDA)
Based on the hypothesis by the client that the rate of customer attrition for their company was driven by price sensitivity wanted to offer customers with high propensity to churn a 20% discount in other to retain them. I was required to build a model to predict customers that are likely to churn. 
 A quick exploratory data analysis (EDA) using "R" to explore(visualize) and clean the messy data was executed followed by the "feature engineering” stage where additional features not in the raw data were computed so that it can be fed into the predictive model that I was about to build. 
  A predictive model was built using the "nnet" package in RStudio to predict customer attrition using the new features from the "feature engineering" stage and some of the cleaned original variables which did not have multicollinearity issues.
 CONCLUSION
A more sophisticated strategy is to find the right level of discount for each customer. This would maximize their predicted revenue or margin.